third_party/blink/web_tests/fast/xsl/preceding-axis-scoped-inside-template-expected.txt

CONSOLE WARNING: crbug.com/435623334: This page uses XSLT, which being considered for removal from the web. If that happens, it is possible that this page will need to be updated to maintain functionality.
This test checks that using the XPath preceding:: axis when instantiating a template (the content of a variable-binding element) does not match nodes outside the template.

2
